MOVED logic to find winner and winner price into AuctionService; UPDATED AUCTION model; CHANGED prices and amounts to be integer type; REMOVED Response and AuctionResponse.

+++ src/Models/Bid.php
@@ -7,15 +7,15 @@ use App\Models\Abstracts\Model;
 
 class Bid extends Model
 {
-    private float $amount;
+    private int $amount;
     private IBuyer $buyer;
 
-    public function setAmount(float $amount) : void
+    public function setAmount(int $amount) : void
     {
         $this->amount = $amount;
     }
 
-    public function getAmount() : float
+    public function getAmount() : int
     {
         return $this->amount;
     }
